// src/openapi/generated/models/resident-chronology-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BedChangeEvent } from '../models/bed-change-event';
import { MeasurementEvent } from '../models/measurement-event';
import { StatusChangeEvent } from '../models/status-change-event';
import { TaskEvent } from '../models/task-event';

/**
 * Respuesta completa de la cronología de un residente
 */
export interface ResidentChronologyResponse {
  events: Array<(MeasurementEvent | TaskEvent | BedChangeEvent | StatusChangeEvent)>;
  resident_id: string;
  resident_name: string;
  total_bed_changes: number;
  total_events: number;
  total_measurements: number;
  total_status_changes: number;
  total_tasks: number;
}