// src/openapi/generated/models/measurement-event.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Evento de medición médica
 */
export interface MeasurementEvent {
  device_name?: (string | null);
  event_type?: 'measurement';
  measurement_id: string;
  measurement_type: string;
  recorded_by?: (string | null);
  recorded_by_name?: (string | null);
  source: string;
  timestamp: string;
  values: {
[key: string]: any;
};
}