// src/openapi/generated/models/task-event.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Evento de tarea asignada/completada
 */
export interface TaskEvent {
  assigned_by?: (string | null);
  assigned_by_name?: (string | null);
  event_type?: 'task';
  recorded_by?: (string | null);
  recorded_by_name?: (string | null);
  status?: (string | null);
  task_application_id: string;
  task_category: string;
  task_name: string;
  timestamp: string;
}