// src/openapi/generated/models/bed-change-event.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Evento de cambio de cama
 */
export interface BedChangeEvent {
  change_type: string;
  event_type?: 'bed_change';
  history_id: number;
  new_bed_id?: (string | null);
  new_location?: (string | null);
  new_room_id?: (string | null);
  previous_bed_id?: (string | null);
  previous_location?: (string | null);
  previous_room_id?: (string | null);
  recorded_by?: (string | null);
  recorded_by_name?: (string | null);
  timestamp: string;
}