// src/openapi/generated/models/status-change-event.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Evento de cambio de estado del residente
 */
export interface StatusChangeEvent {
  event_type?: 'status_change';
  history_id: number;
  new_status: string;
  previous_status?: (string | null);
  recorded_by?: (string | null);
  recorded_by_name?: (string | null);
  timestamp: string;
}